chore: remove deprecated mockData file and decouple component types
